pyproject.toml
[project]
name = "help-bot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "aiofiles==24.1.0",
    "aiogram==3.22.0",
    "aiohappyeyeballs==2.6.1",
    "aiohttp==3.12.15",
    "aiohttp-socks==0.11.0",
    "aiosignal==1.4.0",
    "annotated-types==0.7.0",
    "asyncpg==0.31.0",
    "attrs==25.4.0",
    "certifi==2025.11.12",
    "frozenlist==1.8.0",
    "greenlet==3.2.4",
    "idna==3.11",
    "magic-filter==1.0.12",
    "multidict==6.7.0",
    "propcache==0.4.1",
    "pydantic==2.11.10",
    "pydantic-core==2.33.2",
    "pydantic-settings==2.12.0",
    "python-dotenv==1.2.1",
    "python-socks==2.8.1",
    "sqlalchemy==2.0.44",
    "typing-extensions==4.15.0",
    "typing-inspection==0.4.2",
    "yarl==1.22.0",
    "yoomoney>=2.0.0",
]
